feat: final reviews columns (cluster, assignmentSet)
