+++ client/app/components/FinalReviewsInProgress.vue
@@ -45,7 +45,7 @@
 </template>
 
 <script setup lang="ts">
-import { Anchor, BaseBadge, Icon } from '#components'
+import { Anchor, Icon } from '#components'
 import {
   FlexRender,
   getCoreRowModel,
@@ -55,14 +55,14 @@ import {
   getSortedRowModel,
   type SortingState,
 } from '@tanstack/vue-table'
-import type { QueueItem } from '~/purple_client'
+import type { QueueItem, RpcPerson } from '~/purple_client'
 import { ANCHOR_STYLE } from '~/utils/html'
 import type { HeadingLevel } from '~/utils/html'
-import { groupBy, uniqBy } from 'lodash-es'
 
 type Props = {
   name?: string
   headingLevel?: HeadingLevel
+  people: RpcPerson[]
 }
 
 const props = withDefaults(defineProps<Props>(), { headingLevel: 2 })
@@ -112,18 +112,7 @@ const columns = [
     header: 'Cluster',
     cell: data => {
       const clusterNumber = data.getValue()?.number
-      if (!clusterNumber) {
-        return '-'
-      }
-      return h('span', [
-        h(Anchor, {
-          href: `/clusters/${clusterNumber}`,
-          class: "inline-flex items-center gap-1 text-blue-600"
-        }, () => [
-          h(Icon, { name: "pajamas:group", class: "h-5 w-5" }),
-          clusterNumber
-        ])
-      ])
+      return columnFormatterCluster(clusterNumber)
     },
     sortingFn: 'alphanumeric',
   }),
@@ -133,74 +122,13 @@ const columns = [
       header: 'Assignees',
       cell: (data) => {
         const assignments = data.getValue()
-        if (!assignments) {
-          return 'No assignments'
-        }
-
-        const rfcToBeId = data.row.original.id
-        if (rfcToBeId === undefined) {
-          throw Error(`Internal error: expected queueItem to have id but was ${JSON.stringify(data.row.original)}`)
-        }
-
-        const listItems: VNode[] = []
-
-        const assignmentsByRoles = groupBy(
+        return columnFormatterAssignments({
           assignments,
-          (assignment) => assignment.role
-        )
-
-        const orderedRoles = Object.keys(assignmentsByRoles)
-          .sort((a, b) => a.localeCompare(b, 'en'))
-
-        for (const role of orderedRoles) {
-          const assignmentsOfRole = assignmentsByRoles[role] ?? []
-
-          const redundantAssignmentsOfSamePersonToSameRole = assignmentsOfRole.filter((assignment, _index, arr) => {
-            const { person } = assignment
-            if (person === undefined || person == null) {
-              return false
-            }
-            const firstAssignmentOfPersonToRole = arr.find(arrAssignment => assignment.person && arrAssignment.person && arrAssignment.person === assignment.person)
-            if (!firstAssignmentOfPersonToRole) {
-              console.log(`Couldn't find first assignment for person #${assignment.person} in`, arr)
-              throw Error(`Internal error. Should be able to find first assignment for person #${assignment.person}. See console`)
-            }
-            // the first assignment of person in the list of assignments should always match the current assignment of person
-            // because there shouldn't be duplicate/redundant assignments
-            // but if the id is different then it is a redundant assignment,
-            // so we'll prompt the user to delete them
-            return assignment.id !== firstAssignmentOfPersonToRole.id
-          })
-
-          listItems.push(h('li', { class: 'flex gap-3' }, [
-            h('span',
-              h(BaseBadge, { label: role, class: 'mr-1' })),
-            h('ul', { class: 'flex flex-col gap-2' }, [
-              ...assignmentsOfRole.map(assignment => {
-                const rpcPerson = people.value.find((p) => p.id === assignment.person)
-                return h(Anchor, {
-                  href: rpcPerson ? `/team/${rpcPerson.id}` : undefined,
-                  class: [ANCHOR_STYLE, 'text-sm nowrap']
-                }, () => [
-                  rpcPerson ? rpcPerson.name : pending ? `...` : '(unknown person)',
-                ])
-              }).reduce((acc, item, index, arr) => {
-                // add commas between items
-                const listItemChildren = []
-                listItemChildren.push(item)
-                if (index < arr.length - 1) {
-                  listItemChildren.push(', ')
-                } else {
-                  listItemChildren.push(' ')
-                }
-                const listItem = h('li', listItemChildren)
-                acc.push(listItem)
-                return acc
-              }, [] as (VNode | string)[])]),
-          ]))
-        }
-
-        return h('ul', { class: 'flex flex-col gap-x-1 gap-y-3' }, listItems)
+          rfcToBeId: data.row.original.id,
+          people: props.people,
+          queueItemsIsPending: pending.value,
+          rowForDebug: data.row.original
+        })
       },
       enableSorting: false,
     }
@@ -239,9 +167,5 @@ const table = useVueTable({
   }
 })
 
-const { data: people, status: peopleStatus, error: peopleError } = await useAsyncData(() => api.rpcPersonList(), {
-  server: false,
-  lazy: true,
-  default: () => [] as RpcPerson[]
-})
+
 </script>
